src/diagrams_generator/requirements.txt
cssutils==2.6.0
Pillow==9.2.0
